added tests put_v1_account_login,  put_v1_account_token and updated test put_v1_account_email
