+++ tests/functional/put_v1_account/test_put_v1_account_email.py
@@ -6,13 +6,14 @@ from dm_api_account.apis.account_api import AccountApi
 from api_mailhog.apis.mailhog_api import MailhogApi
 
 
-def test_put_v1_account():
+def test_put_v1_account_email():
     account_api = AccountApi(host='http://5.63.153.31:5051')
     login_api = LoginApi(host='http://5.63.153.31:5051')
     mailhog_api = MailhogApi(host='http://5.63.153.31:5025')
 
-    login = 'IM_test_13'
+    login = 'IM_test_e9'
     email = f'{login}@mail.com'
+    email_upd = f'{login}_upd@mail.com'
     password = 'pass123456'
 
     #  -------User registration
@@ -26,7 +27,7 @@ def test_put_v1_account():
     response = account_api.post_v1_account(json_data=json_data)
     print(response.status_code, '  - User creation status code')
     print(response.text)
-    assert response.status_code == 201, f"User wasn't created {response.json()}"
+    # assert response.status_code == 201, f"User wasn't created {response.json()}"
 
     # -------Get mail from mail server
 
@@ -37,6 +38,7 @@ def test_put_v1_account():
     #  -------Get activation token
 
     token = get_activation_token_by_login(login, response)
+
     assert token is not None, f"Token not received for user {login}"
 
     #  -------Activate user
@@ -58,6 +60,8 @@ def test_put_v1_account():
     print(response.status_code, ' - Login status code for', login)
     print(response.text)
     assert response.status_code == 200, "User didn't logged in"
+
+    # -------------------------------------------------------------------------
 
     #  -------Update email
 
@@ -87,10 +91,32 @@ def test_put_v1_account():
     print(response.text)
     assert response.status_code != 200, "User  logged in successfully"
 
+    # -------Get mail from mail server
+
+    response = mailhog_api.get_api_v2_messages()
+    print(response.status_code, ' - Get mails status code')
+    assert response.status_code == 200, "Mails not received"
+
+    #  -------Get activation token after changing mail
+    # email_upd = f'{login}_upd@mail.com',
+
+    json_data = {
+        'login': login,
+        'email_upd': f'{login}_upd@mail.com',
+        'password': password,
+    }
+
+    token = get_activation_token_by_login_after_changing_mail(
+        login,
+        email_upd,
+        response
+    )
+
     #  -------Activate user after changing email
     response = account_api.put_v1_account_token(token)
 
-    print(response.status_code, ' - User activation status code')
+    print(response.status_code, ' - User', login, ' activation status code')
+
     assert response.status_code == 200, "User not activated"
 
     # ---Log in 3 user after changing email and new activation
@@ -117,4 +143,23 @@ def get_activation_token_by_login(
         if user_login == login:
             token = user_data['ConfirmationLinkUrl'].split('/')[-1]
             print(login, token)
+    return token
+
+
+def get_activation_token_by_login_after_changing_mail(
+        login,
+        email_upd,
+        response
+):
+    for item in response.json()['items']:
+        mail_to = (item['Raw']['To'][0])
+        if mail_to == email_upd:
+            user_data = (loads(item['Content']['Body']))
+            user_data = (loads(item['Content']['Body']))
+            user_login = user_data['Login']
+
+            if user_login == login:
+                token = user_data['ConfirmationLinkUrl'].split('/')[-1]
+                print('for user=', login,  'user current mail=', mail_to, 'token=' , token,)
+                break
     return token
